allow base URL to have same value as upstream.
sometimes you do not want a different domain for the auth/registration
views of the webapp. in this cases you can use the same URL as one of
the upstream services has. doorman now uses a specific URL parameter;
this will clash with an app if the app has the sampe parameter (very
unlikely). but it you have the chance: use a different domain for the
issuer_base.